# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a modern RSS reader API.",
        "title": "Gist API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/ai/summarize": {
            "post": {
                "description": "Generate an AI summary of the article content. Returns cached result if available, otherwise streams the response.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "Generate AI summary",
                "parameters": [
                    {
                        "description": "Summarize request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.summarizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cached summary",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.summarizeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/ai/translate": {
            "post": {
                "description": "Translate article content. Returns cached result if available, otherwise streams block translations via SSE.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "Generate AI translation",
                "parameters": [
                    {
                        "description": "Translate request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.translateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.translateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/ai/translate/batch": {
            "post": {
                "description": "Translate multiple articles' titles and summaries. Returns NDJSON stream.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/x-ndjson"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "Batch translate articles",
                "parameters": [
                    {
                        "description": "Batch translate request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.batchTranslateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gist_backend_internal_service.BatchTranslateResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gist_backend_internal_service.BatchTranslateResult": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.batchTranslateRequest": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "summary": {
                                "type": "string"
                            },
                            "title": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "internal_handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.summarizeRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "entryId": {
                    "type": "string"
                },
                "isReadability": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_handler.summarizeResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "internal_handler.translateRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "entryId": {
                    "type": "string"
                },
                "isReadability": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_handler.translateResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean"
                },
                "content": {
                    "type": "string"
                }
            }
        },
...
    }
}